+++ src/hosted-fields/internal/models/credit-card-form.js
@@ -214,18 +214,7 @@ CreditCardForm.prototype._onNumberChange = function (number, metadata) {
     (oldNumberIsShortEnoughForBinEvent || oldBinIsNotEqualToNewBin) &&
     newNumberIsLongEnoughForBinEvent
   ) {
-    // TODO SVM-3806
-    // Temporary fix for SVM-3806
-    // Once a more comprehensive solution for targetting
-    // only specific frames is introduced, remove this
-    if (document.referrer) {
-      window.bus
-        .target({
-          channel: window.bus.channel,
-          origin: document.referrer.replace(/(^\/|\/$)/g, ""),
-        })
-        .emit(events.BIN_AVAILABLE, newBin);
-    }
+    window.bus.emit(events.BIN_AVAILABLE, newBin);
   }
 };